+++ Taskfile.yml
@@ -15,7 +15,7 @@ tasks:
 
   test:
     cmds:
-      - kind delete cluster --name=tests && kind create cluster --name=tests
+      # - kind delete cluster --name=tests && kind create cluster --name=tests
       - go test -coverprofile cover.out -p 1 -v ./...
 
   update-deps:
