+++ cmd/yoke/main_test.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 	"io"
 	"math/rand/v2"
+	"os"
 	"strconv"
 	"strings"
 	"testing"
@@ -26,8 +27,22 @@ import (
 	"github.com/yokecd/yoke/internal"
 	"github.com/yokecd/yoke/internal/home"
 	"github.com/yokecd/yoke/internal/k8s"
+	"github.com/yokecd/yoke/internal/testutils"
 	"github.com/yokecd/yoke/pkg/yoke"
 )
+
+func TestMain(m *testing.M) {
+	must := func(err error) {
+		if err != nil {
+			panic(err)
+		}
+	}
+
+	must(testutils.X("kind delete clusters --all"))
+	must(testutils.X("kind create cluster --name=tests"))
+
+	os.Exit(m.Run())
+}
 
 var background = context.Background()
